requirements.txt
fastapi==0.135.3
uvicorn==0.43.0
sqlalchemy==2.0.49
psycopg2-binary==2.9.11
python-dotenv
pydantic==2.12.5
celery==5.3.4
redis==7.4.0
requests
flower==2.0.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2
httpx==0.25.2